:truck: Rename to categorical series widget
