+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/countCategorical.tsx
