+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/durationCategorical.tsx
