+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/largeValueCategorical.tsx
