+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/longLabelCategorical.tsx
