+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/manyCategoriesCategorical.tsx
