+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/multiSeriesCategorical.tsx
