+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/negativeCategorical.tsx
