+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/percentageCategorical.tsx
