+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/rateCategorical.tsx
