+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/sizeCategorical.tsx
