+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/fixtures/sparseCategorical.tsx
