+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/formatters/formatCategoricalSeriesName.tsx
