+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/plottables/bars.spec.tsx
@@ -1,9 +1,9 @@
 import {CategoricalSeriesFixture} from 'sentry-fixture/categoricalSeries';
 
-import {sampleCountCategoricalData} from 'sentry/views/dashboards/widgets/barChartWidgetVisualization/fixtures/countCategorical';
-import {sampleDurationCategoricalData} from 'sentry/views/dashboards/widgets/barChartWidgetVisualization/fixtures/durationCategorical';
-import {sampleStackedCategoricalData} from 'sentry/views/dashboards/widgets/barChartWidgetVisualization/fixtures/stackedCategorical';
-import {Bars} from 'sentry/views/dashboards/widgets/barChartWidgetVisualization/plottables/bars';
+import {sampleCountCategoricalData} from 'sentry/views/dashboards/widgets/categoricalSeriesWidget/fixtures/countCategorical';
+import {sampleDurationCategoricalData} from 'sentry/views/dashboards/widgets/categoricalSeriesWidget/fixtures/durationCategorical';
+import {sampleStackedCategoricalData} from 'sentry/views/dashboards/widgets/categoricalSeriesWidget/fixtures/stackedCategorical';
+import {Bars} from 'sentry/views/dashboards/widgets/categoricalSeriesWidget/plottables/bars';
 
 describe('Bars', () => {
   describe('Properties', () => {
