+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/plottables/bars.tsx
