+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/plottables/plottable.tsx
