+++ static/app/views/dashboards/widgets/categoricalSeriesWidget/settings.tsx
